Implement runnable JAR file option

+++ run.sh
@@ -24,20 +24,31 @@ then
     listDirectoryFilter=$(echo $listDirectory | cut -f 1 -d '/')
     
     extension=$(echo $1 | cut -f 1 -d '.')
-    echo "$PKG"
     if [ "$listDirectoryFilter" = "$PKG" ]
     then
-        echo "Your package directory is $PKG.$extension"
+        echo -e "\nYour package directory is $PKG.$extension"
         echo "PROCEED? Y/n"
         read Input
         if [ "$Input" == "Y" ]
         then
             #echo "Run java command on terminal: java -cp . $PKG,$extension"
-            echo "Run bash start.sh on terminal"
-            echo "java.exe -cp . $PKG.$extension" > start.sh
-            exit
+            
+            echo -e "\n=======QUICK DETOUR======"
+            echo -e "\nDo you like to explore option for creating a runnable JAR file [Y/n]"
+            read option
+            if [ "$option" == "Y" ]
+            then
+                echo "echo -e '\nCreating JAR file!\n'" > startJar.sh
+                echo "jar.exe cvmf manifest.txt test.jar $listDirectory$extension.class" >> startJar.sh
+                echo "java.exe -jar test.jar" >> startJar.sh
+                #$(jar.exe cvmf "manifest.txt" "test.jar" $listDirectory$extension.class)
+                echo -e "\nRun 'bash startJar.sh' on the terminal."
+            else
+                echo "java.exe -cp . $PKG.$extension" > start.sh
+                echo "Run bash start.sh on terminal."
+            fi
         else
-            echo "Please check your package syntax."
+            echo "Thank you for your visit!"
         fi
     else
         echo "Check your directory listings!"
